+++ opentelemetry-sdk/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixed
 
 - Fix metric export corruption if gauges have not received a last value. (#1363)
+- Return consistent `Meter` for a given scope from `MeterProvider`. (#1351)
 
 ## v0.21.0